Refactor duplicated ogg code used by ogg vorbis and ogg opus
